pipeline/requirements.txt
beautifulsoup4==4.11.1
feedparser==6.0.10
azure-core==1.20.1
azure-storage-blob==12.9.0
click==8.0.3
pandas==1.2.4
python-dotenv==0.21.1
requests==2.28.2
pycurl==7.45.2